2024-08-11 018: [DeduceTypes2] Split into separate classes
1. This drives me crazy
2. Too big anyway (confuses "some" tools, understandably)
