+++ elevated-potential.core.impl/src/main/java/tripleo/elijah/stages/deduce/ExpectationBase.java
@@ -1,0 +1,5 @@
+package tripleo.elijah.stages.deduce;
+
+public interface ExpectationBase {
+	String expectationString();
+}
