+++ elevated-potential.core.impl/src/main/java/tripleo/elijah/stages/deduce/IElementProcessor.java
@@ -1,0 +1,9 @@
+package tripleo.elijah.stages.deduce;
+
+import tripleo.elijah.lang.i.*;
+
+public interface IElementProcessor {
+	void elementIsNull();
+
+	void hasElement(OS_Element el);
+}
