+++ elevated-potential.core.impl/src/main/java/tripleo/elijah/stages/deduce/IVariableConnector.java
@@ -1,0 +1,7 @@
+package tripleo.elijah.stages.deduce;
+
+import tripleo.elijah.stages.gen_fn.*;
+
+interface IVariableConnector {
+	void connect(VariableTableEntry aVte, String aName);
+}
