+++ elevated-potential.core.impl/src/main/java/tripleo/elijah/stages/deduce/NullConnector.java
@@ -1,0 +1,9 @@
+package tripleo.elijah.stages.deduce;
+
+import tripleo.elijah.stages.gen_fn.*;
+
+class NullConnector implements IVariableConnector {
+	@Override
+	public void connect(final VariableTableEntry aVte, final String aName) {
+	}
+}
